% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_fossils_track_node.R
\name{drop_fossils_track_node}
\alias{drop_fossils_track_node}
\title{Drop fossil tips and record matching nodes}
\usage{
drop_fossils_track_node(phy)
}
\arguments{
\item{phy}{a phylogeny with fossil tips}
}
\value{
a list with the pruned phylogeny and a table with the matching nodes
}
\description{
Drop fossil tips and record matching nodes
}
