% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMK_budding.R
\name{make_scaler_simmap}
\alias{make_scaler_simmap}
\title{Make stochastic map with rate scalers}
\usage{
make_scaler_simmap(sims, ncat = 10, low_cut, high_cut)
}
\arguments{
\item{sims}{the simulation object output from 'sim_Mk_budding_exp'.}

\item{ncat}{number of categories to make the legend for the scalers plot.}

\item{low_cut}{break for the first category (slow rate scalers). Rate scaler values above 'low_cut' value will be divided into 'ncat'-1 categories.}

\item{high_cut}{break for the last category (fast rate scalers). Rate scaler values below 'high_cut' value will be divided into 'ncat'-1 categories.}
}
\value{
a list with a simmap object and a matrix with legend information
}
\description{
Plot the rate scalers used in the simulation to the phylogenetic tree. Also return a vector of colors to be used for visualization of the gradient paired with the breaking points for the rate scaler values. \cr
Use 'low_cut' or 'high_cut' to focus the rate categories into the slow or fast rate categories. This can be used to provide a more informative plot of rates along the branches of the phylogeny.
}
