% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cont_sim.R
\name{plot_cont_sim}
\alias{plot_cont_sim}
\title{Plot trajectory of continuous trait simulation}
\usage{
plot_cont_sim(
  sim_obj,
  edge_width = 1.5,
  bg = "white",
  jump_col = "coral",
  jump_width = 1,
  main = "",
  legend = TRUE
)
}
\arguments{
\item{sim_obj}{the output of a 'buddPhy' simulation function.}

\item{edge_width}{the width of the lines}

\item{bg}{the color for the background of the plot. Helps to visualize some colors. Use "white" to turn it off.}

\item{jump_col}{the color of the line connecting the states for the nodes in which a phenotype jump happened after budding speciation.}

\item{jump_width}{the line width for the jump lines.}

\item{main}{a title for the plot.}

\item{legend}{if to add a legend. Note that legend can get in the way of the plot sometimes.}
}
\value{

}
\description{
Function plots the trajectory of a single continuous trait evolving along the branches of the tree following a lineage-age dependent process including events of budding (or symmetric) speciation.
}
