% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMK_budding.R
\name{plot_mother_lineages}
\alias{plot_mother_lineages}
\title{Plot phylogeny and show mother lineages}
\usage{
plot_mother_lineages(
  sim_obj,
  background_color = "gray",
  edge_width = 3,
  no_margin = TRUE
)
}
\arguments{
\item{sim_obj}{the output of a buddPhy simulation function.}

\item{background_color}{the color of the non-mother lineages.}

\item{edge_width}{the width of the phylogeny branches.}

\item{no_margin}{if to add margin. See 'ape::plot.phylo'.}
}
\value{
Plots a phylogenetic tree.
}
\description{
Function plots the phylogenetic tree from a buddPhy simulation and highlights the mother lineages. \cr
Mother lineages are those that survived through one of more events of speciation via budding.
}
