% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMK_trace_history.R
\name{sim_Mk_trace_history}
\alias{sim_Mk_trace_history}
\title{Simulate discrete trait evolution tracing lineage history.}
\usage{
sim_Mk_trace_history(
  sim_MK,
  Q,
  anc = NULL,
  change_rate = 2,
  decay_fn = TRUE,
  cladogenetic_change = "none"
)
}
\arguments{
\item{sim_MK}{the output from ''sim_Mk_budding_exp'.}

\item{Q}{transition matrix for a Markov model.}

\item{anc}{state at the root. If "NULL", then a random state is selected.}

\item{change_rate}{the rate for an exponential reduction of the rate of trait evolution proportional to lineage-age.}

\item{decay_fn}{TRUE or FALSE. If TRUE, then the rate of trait evolution has a negative (exponential) relationship with lineage-age. If FALSE, then this relationship becomes positive.}

\item{cladogenetic_change}{if a trait change should occur in the node associated with a budding event. Option are "none": no change, anagenetic changes only; "flat": trait changes randomly; "prob": trait changes following the transition probabilities from the Q matrix.}
}
\value{
A list with simmap, tip_state, edge_state, ancestry, rate_scaler, and scaler_mat.
}
\description{
This function simulates discrete traits based on the same lineage history created when using 'sim_Mk_budding_exp'. It is useful to simulate under the same lineage history and phylogenetic tree while changing rate parameters or parameter for the decay function. If you don't want to retain a particular lineage history, then please refer to 'sim_Mk_budding_exp'.
}
