% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_estimation.R
\name{tree_rescale_exp}
\alias{tree_rescale_exp}
\title{Rescale branch lengths following the model}
\usage{
tree_rescale_exp(
  change_rate = 0.05,
  budding_prob = 0.3,
  tree,
  chunk_fraction = 0.01,
  decay_fn = TRUE
)
}
\arguments{
\item{change_rate}{rate for the exponential distribution.}

\item{budding_prob}{probability of a budding speciation on the internal node.}

\item{tree}{phylogenetic tree.}

\item{chunk_fraction}{fraction of the total tree-height for branch length discretization.}

\item{decay_fn}{whether rates should vary proportional (FALSE) or inversely proportional (TRUE) to lineage age.}
}
\value{
a phylogenetic tree with rescaled branch lengths and reordered to "cladewise" order (see 'ape::reorder.phylo').
}
\description{
Function to rescale branches of a phylogenetic tree following the exponential rate proportional, or inversely-proportional, to lineage-age. This function can be used to simulate a realization of the model for posterior predictive checks among other reasons.
}
