% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMK_budding.R
\name{get_Budding_History}
\alias{get_Budding_History}
\title{Extract budding history from simulations}
\usage{
get_Budding_History(sim_Mk)
}
\arguments{
\item{sim_Mk}{A simulation object generated with 'sim_Mk_budding_exp'}
}
\value{
a list. budd_nodes are the nodes representing budding speciation events. budd_edges is an index, in the same order of phy$edge, selecting all branches belonging to mother lineages. A mother lineage is a lineage that survived one or more speciation events (because of budding).
}
\description{
This function can be used to extract the history of budding events across the branches and nodes of the tree.
}
