% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBM_budding.R
\name{sim_BM_budding_exp}
\alias{sim_BM_budding_exp}
\title{Simulate continuous traits under a lineage-age dependent model}
\usage{
sim_BM_budding_exp(
  tree,
  sigma,
  anc = 0,
  budding_prob = 0,
  budding_mother = NA,
  change_rate = 2,
  decay_fn = TRUE
)
}
\arguments{
\item{tree}{phylogeny with branch lengths.}

\item{sigma}{the base rate for the Brownian-motion model.}

\item{anc}{the ancestral state for the trait.}

\item{budding_prob}{probability that the speciation event was budding.}

\item{budding_mother}{autocorrelation of budding events. If NA, then no autocorrelation happens and the occurrence of budding speciation will always follow the same probability of 'budding_prob'. If 'budding_mother' is non-zero, then the probability that the mother lineage will generate another new lineage through a budding speciation event (and, thus, survive another speciation event) will be given by 'budding_mother' instead. The probability of the next budding event after a symmetrical speciation event will reset to 'budding_prob'.}

\item{change_rate}{the rate for an exponential reduction of the rate of trait evolution proportional to lineage-age.}

\item{decay_fn}{TRUE or FALSE. If TRUE, then the rate of trait evolution has a negative (exponential) relationship with lineage-age. If FALSE, then this relationship becomes positive.}
}
\value{
A list with the information from the simulation and the states at the tips.
}
\description{
Simulate continuous traits under a lineage-age dependent model
}
