% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBM_budding.R
\name{sim_BM_trace_history}
\alias{sim_BM_trace_history}
\title{Simulate continuous trait evolution tracing lineage history.}
\usage{
sim_BM_trace_history(
  sim_BM,
  sigma,
  anc = 0,
  change_rate = 2,
  decay_fn = TRUE,
  jump_size = NA
)
}
\arguments{
\item{sim_BM}{output from the 'sim_BM_budding_exp'.}

\item{sigma}{the base rate for the Brownian-motion model.}

\item{anc}{the ancestral state for the trait.}

\item{change_rate}{the rate for an exponential reduction of the rate of trait evolution proportional to lineage-age.}

\item{decay_fn}{TRUE or FALSE. If TRUE, then the rate of trait evolution has a negative (exponential) relationship with lineage-age. If FALSE, then this relationship becomes positive.}

\item{jump_size}{controls the size of the cladogenetic change associated with budding speciation. The new lineage will have an average trait value shifted by a proportion equal to 'jump_size'. For example, 'jump_size = 0.5' will cause the new lineages to start with an average trait value equal to 50\% above or below the current trait value. Setting to NA (default) will turn-off the option of cladogenetic trait changes.}
}
\value{
A list with the information from the simulation and the states at the tips.
}
\description{
This function simulates continuous traits based on the same lineage history created when using 'sim_BM_budding_exp'. It is useful to simulate under the same lineage history and phylogenetic tree while changing rate parameters or parameter for the decay function. If you don't want to retain a particular lineage history, then please refer to 'sim_BM_budding_exp'.
}
