% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMK_budding.R
\name{sim_Mk_budding_exp_decay}
\alias{sim_Mk_budding_exp_decay}
\title{Simulate discrete trait evolution with budding and exponential decay function.}
\usage{
sim_Mk_budding_exp_decay(
  tree,
  Q,
  anc = NULL,
  budding_prob = 0,
  decay_rate = 2,
  cladogenetic_change = FALSE
)
}
\arguments{
\item{tree}{phylogeny with branch lengths.}

\item{Q}{transition matrix for a Markov model.}

\item{anc}{state at the root. If "NULL", then a random state is selected.}

\item{budding_prob}{probability that the speciation event was budding.}

\item{decay_rate}{the rate for an exponential reduction of the rate of trait evolution proportional to lineage-age.}

\item{cladogenetic_change}{if the state of the budding lineage should be distinct from the ancestral lineage (TRUE or FALSE).}
}
\value{
A list with simmap, tip_state, edge_state, ancestry, rate_scaler, and scaler_mat.
}
\description{
This function adds a probability of budding speciation and allows only for a negative relationship between rates of trait evolution and lineage-age. Function also allows for cladogenetic change in the trait (a deterministic change in the state of the trait after budding speciation happens).
}
